fledgeForGpt: delay slot config to end of auction
